// src/pages/modules/E-Office/Letter-Generation/Letter-type/detail-template/data/types.ts
export interface IMailTypeLetterTemplate {
  id_mail_template_surat: string
  id_satuan_organisasi: string
  id_jenis_template_surat: string
  urutan: number
  uraian: string
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_user_created: string
  nama_user_updated: string
  nama_jenis_template: string
}